Update list and api
