+++ pages/list.tsx
@@ -1,8 +1,10 @@
-import { Container } from "ooni-components"
-
+import ListLoader from "components/ListLoader"
+import { getList } from "lib/api"
 import dynamic from "next/dynamic"
-import { Box, Heading } from "ooni-components"
+import { Box, Container, Heading } from "ooni-components"
+import { useMemo } from "react"
 import styled from "styled-components"
+import useSWR from "swr"
 
 const RunLinkList = dynamic(() => import("components/List"))
 const OONIRunHero = dynamic(() => import("components/OONIRunHero"))
@@ -14,6 +16,20 @@ box-shadow: 0 50vh 0 50vh ${(props) => props.theme.colors.gray1};
 const List = () => {
   const queryParams = { only_mine: true, include_archived: true }
 
+  const { data, error, isLoading } = useSWR<{ descriptors: Descriptor[] }>(
+    { only_latest: true, ...queryParams },
+    (props: object) => getList(props),
+  )
+
+  const descriptors = useMemo(
+    // archived links are shown at the end
+    () =>
+      data?.descriptors.sort(
+        (a, b) => Number(a.archived) - Number(b.archived),
+      ) || [],
+    [data],
+  )
+
   return (
     <>
       <OONIRunHero />
@@ -22,7 +38,8 @@ const List = () => {
           <Heading h={2} mb={2}>
             My OONI Run Links
           </Heading>
-          <RunLinkList queryParams={queryParams} />
+          {isLoading && <ListLoader />}
+          {!!descriptors.length && <RunLinkList descriptors={descriptors} />}
         </Container>
       </StyledBox>
     </>
